Move Image to module, update filename
