+++ src/backend/Positive/Server.hs
@@ -23,15 +23,14 @@ import qualified Data.ByteString.Lazy as ByteString
 import qualified Data.Text as Text
 import qualified Data.Time.Clock as Time
 import qualified Data.Vector.Unboxed as Vector
-import GHC.Float (int2Double)
 import qualified Graphics.Image as HIP
-import qualified Network.HTTP.Media as Media
 import Network.Wai.Handler.Warp
 import Positive.Effect.FilmRoll as FilmRoll
 import Positive.Effect.Log
 import Positive.Flags
+import Positive.Image
+import Positive.ImageSettings as ImageSettings
 import Positive.Prelude hiding (ByteString)
-import Positive.Settings as Settings
 import qualified Positive.Static as Static
 import Servant hiding (throwError)
 import Servant.API.Generic
@@ -168,11 +167,11 @@ handleSaveSettings ::
   m [ImageSettings]
 handleSaveSettings settings = do
   Env {ePreviewMVar} <- ask
-  let newSettings = Settings.fromList settings
+  let newSettings = ImageSettings.fromList settings
   FilmRoll.writeSettings newSettings
   log "Updated settings"
   sendIO $ pSwapMVar ePreviewMVar newSettings
-  pure $ Settings.toList newSettings
+  pure $ ImageSettings.toList newSettings
 
 handleGetSettings ::
   ( Has (Reader Env) sig m,
@@ -180,7 +179,7 @@ handleGetSettings ::
   ) =>
   m ([ImageSettings], PathSegment)
 handleGetSettings =
-  (\a b -> (Settings.toList a, eDir b)) <$> FilmRoll.readSettings <*> ask
+  (\a b -> (ImageSettings.toList a, eDir b)) <$> FilmRoll.readSettings <*> ask
 
 handleGenerateHighRes ::
   ( Has (Lift IO) sig m,
@@ -247,75 +246,6 @@ getImage previewWidth path = do
       log "Read image"
       pure (image, putMVar eImageMVar (path, image))
 
--- IMAGE
-
-type MonochromeImage arr =
-  HIP.Image arr HIP.Y Double
-
-type MonochromePixel =
-  HIP.Pixel HIP.Y Double
-
-readImageFromDisk :: String -> IO (MonochromeImage HIP.VU)
-readImageFromDisk =
-  HIP.readImageY HIP.VU
-
-resizeImage :: Int -> MonochromeImage HIP.VU -> MonochromeImage HIP.VU
-resizeImage targetWidth image =
-  let mul = int2Double (HIP.rows image) / int2Double (HIP.cols image)
-   in HIP.resize
-        (HIP.Bicubic (-0.25))
-        HIP.Edge
-        (round (int2Double targetWidth * mul), targetWidth)
-        image
-
-processImage :: ImageSettings -> MonochromeImage HIP.VU -> MonochromeImage HIP.VU
-processImage is image =
-  let (h, w) = HIP.dims image
-      cropOffset =
-        ( floor $ int2Double h / 100 * icTop (iCrop is),
-          floor $ int2Double w / 100 * icLeft (iCrop is)
-        )
-      cropWidth = floor $ int2Double (w - snd cropOffset) * (icWidth (iCrop is) / 100)
-      cropHeight = floor $ int2Double cropWidth * mul
-      mul = int2Double h / int2Double w
-   in HIP.map
-        ( zone 0.9 (iZone9 is)
-            . zone 0.5 (iZone5 is)
-            . zone 0.1 (iZone1 is)
-            . gamma (iGamma is)
-            . compress (iBlackpoint is) (iWhitepoint is)
-            . invert
-        )
-        -- NOTE: 'normalize' is the only "automatic" correction and has a subtle effect
-        -- at which it makes smaller images have more contrast due to different max and min
-        -- values vs the original size
-        . HIP.normalize
-        . (if iRotate is == 0 then id else HIP.rotate (HIP.Bicubic (-0.25)) (HIP.Fill 0) (iRotate is))
-        $ HIP.crop cropOffset (cropHeight, cropWidth) image
-
--- FILTERS
-
-invert :: MonochromePixel -> MonochromePixel
-invert =
-  fmap (1 -)
-{-# INLINE invert #-}
-
-compress :: Double -> Double -> MonochromePixel -> MonochromePixel
-compress s l =
-  fmap (\p -> min 1 . max 0 $ (p - s) / (l - s))
-{-# INLINE compress #-}
-
-gamma :: Double -> MonochromePixel -> MonochromePixel
-gamma x =
-  fmap (** x)
-{-# INLINE gamma #-}
-
-zone :: Double -> Double -> MonochromePixel -> MonochromePixel
-zone t i =
-  let m v = (1 - abs (v - t)) ^ (2 :: Int)
-   in fmap (\v -> v + (i * m v))
-{-# INLINE zone #-}
-
 -- PREVIEW LOOP
 
 previewWorker :: TimedFastLogger -> MVar FilmRollSettings -> PathSegment -> IO ()
@@ -343,17 +273,6 @@ previewWorker logger previewMVar dir =
               =<< HIP.encode HIP.JPG [] . HIP.exchange HIP.VS . processImage settings . resizeImage 750
               <$> readImageFromDisk (segmentToString input)
         threadDelay 10000000
-
--- IMAGE
-
-data Image
-
-instance Accept Image where
-  contentType _ =
-    "image" Media.// "jpg"
-
-instance MimeRender Image ByteString where
-  mimeRender _ = id
 
 -- PROFILE
